requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=0.30.1=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
biopython=1.79=py39hb9d737c_3
blast=2.5.0=hc0b0e79_3
boost=1.78.0=py39hac2352c_0
boost-cpp=1.78.0=h75c5d50_1
bzip2=1.0.8=hd590300_5
ca-certificates=2025.1.31=hbcca054_0
certifi=2024.6.2=pypi_0
cgecore=1.5.6=pypi_0
cgelib=0.7.5=pypi_0
charset-normalizer=3.3.2=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
datasets=2.19.2=pypi_0
dill=0.3.8=pypi_0
einops=0.8.0=pyhd8ed1ab_1
filelock=3.14.0=pypi_0
fonttools=4.53.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
huggingface-hub=0.23.3=pypi_0
icu=70.1=h27087fc_0
idna=3.7=pypi_0
importlib-resources=6.4.0=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.40=hf3520f5_2
libblas=3.9.0=8_openblas
libcblas=3.9.0=8_openblas
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h77fa898_7
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libgomp=13.2.0=h77fa898_7
liblapack=3.9.0=8_openblas
libnsl=2.0.1=hd590300_0
libopenblas=0.3.12=pthreads_hb3c22a3_1
libsqlite=3.45.3=h2797004_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=h4ab18f5_1
lz4-c=1.9.3=h9c3ff4c_1
markupsafe=2.1.5=pypi_0
matplotlib=3.9.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.5=h59595ed_0
networkx=3.2.1=pypi_0
numpy=2.0.2=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.5.40=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openssl=3.3.1=hb9d3cd8_3
orffinder=1.8=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
peft=0.11.1=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=pyhd8ed1ab_0
prodigal=2.6.3=h031d066_9
psutil=6.0.0=pypi_0
pyarrow=16.1.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyparsing=3.1.2=pypi_0
python=3.9.19=h0755675_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.9=5_cp39
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h8228510_1
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
resfinder=4.6.0=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=70.0.0=pyhd8ed1ab_0
six=1.16.0=pypi_0
smmap=5.0.2=pypi_0
sympy=1.12.1=pypi_0
tabulate=0.9.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.19.1=pypi_0
torch=2.3.1=pypi_0
tqdm=4.66.4=pypi_0
transformers=4.41.2=pypi_0
triton=2.3.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xxhash=3.4.1=pypi_0
xz=5.2.6=h166bdaf_0
yarl=1.9.4=pypi_0
zipp=3.20.2=pypi_0
zstd=1.5.2=h8a70e8d_1
